resolve PR changes, homebtn forms, %tag
