+++ src/components/pages/ContactSuccess.jsx
@@ -1,7 +1,13 @@
+import { Link } from "react-router-dom";
+import { StyledButton } from "../../style";
+
 export default function ContactSuccess() {
   return (
     <>
       <h1>Your message has been sent!</h1>
+      <Link to="/">
+        <StyledButton>Take me home</StyledButton>
+      </Link>
     </>
   );
 }
